demo/persistency/requirements.txt
PyYAML==3.10
pystache==0.3.1
